LayoutTests/fast/regions/webkit-region-overflow-parsing-expected.txt

Tests parsing of webkit-region-overflow property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testCSSText("-webkit-region-overflow: auto") is "auto"
PASS testCSSText("-webkit-region-overflow: initial") is "initial"
PASS testCSSText("-webkit-region-overflow: inherit") is "inherit"
PASS testCSSText("-webkit-region-overflow: break") is "break"
PASS testCSSText("-webkit-region-overflow: 0") is ""
PASS testCSSText("-webkit-region-overflow: -1") is ""
PASS testCSSText("-webkit-region-overflow: 12.5") is ""
PASS testCSSText("-webkit-region-overflow: 1px") is ""
PASS testComputedStyle("auto") is "auto"
PASS testComputedStyle("initial") is "auto"
PASS testComputedStyle("inherit") is "auto"
PASS testComputedStyle("break") is "break"
PASS testComputedStyle("0") is "auto"
PASS testComputedStyle("-1") is "auto"
PASS testComputedStyle("12.5") is "auto"
PASS testComputedStyle("1px") is "auto"
PASS successfullyParsed is true

TEST COMPLETE
